Added GPUBufferedArray and updated gpucodegen, moved to brian.experimental.cuda
GPUBufferedArray can be used for mixing GPU aware code with code that isn't. It's an object that keeps a copy of the same data on the GPU and CPU and before the data on one is changed, it is synchronised if necessary (it doesn't always work, but in most cases it will).

gpucodegen module has been updated to use the buffering mechanism.

Both have been moved to the new brian.experimental.cuda package.