Sparse connection matrix init improvements
Now uses weave if available, or mergesort based algorithm if not. mergesort algorithm is in principle O(nnz log nnz) compared to pure python algorithm which was O(nnz) but in practice for the number of synapses at the memory limit (close to 2gb) the mergesort algorithm is still much faster (and of almost exactly the same time as the weave based one).