Refractoriness fix (temporary)
New mechanism disallows spikes for period refractory, but involves two extra lines in the update() method of NeuronGroup which may be costly (untested yet).

This fix may also introduce other bugs as the _next_allowed_spiketime array should be reset to 0 before rerunning which happens if NeuronGroup.reinit() is called, but this may not always be the case.