New Brian hears: multiple changes to Sound class following the notes in meeting.txt
Most importantly, there is now a dB object (of type dB_type) which behaves a bit like a Quantity and raises errors if you try to do non-dB things with it. There is also a gain(level) function, and an exception dB_error.

The intensity functions have been replaced by the level functions, including the level computed attribute. In single channel sounds, dB type objects have to be given for levels, but for multi channel sounds arrays are returned so these are not dB type objects.